requirements.txt
fastapi
uvicorn
astrapy
python-dotenv
passlib[bcrypt]
fastapi-users